# [pdata] Introduce runtime safeguards to catch incorrect pdata mutations