--- a/pdata/internal/generated_wrapper_float64slice.go
+++ b/pdata/internal/generated_wrapper_float64slice.go
@@ -7,13 +7,18 @@
 package internal
 
 type Float64Slice struct {
-	orig *[]float64
+	orig  *[]float64
+	state *State
 }
 
 func GetOrigFloat64Slice(ms Float64Slice) *[]float64 {
 	return ms.orig
 }
 
-func NewFloat64Slice(orig *[]float64) Float64Slice {
-	return Float64Slice{orig: orig}
+func GetFloat64SliceState(ms Float64Slice) *State {
+	return ms.state
+}
+
+func NewFloat64Slice(orig *[]float64, state *State) Float64Slice {
+	return Float64Slice{orig: orig, state: state}
 }