--- a/pdata/internal/generated_wrapper_uint64slice.go
+++ b/pdata/internal/generated_wrapper_uint64slice.go
@@ -7,13 +7,18 @@
 package internal
 
 type UInt64Slice struct {
-	orig *[]uint64
+	orig  *[]uint64
+	state *State
 }
 
 func GetOrigUInt64Slice(ms UInt64Slice) *[]uint64 {
 	return ms.orig
 }
 
-func NewUInt64Slice(orig *[]uint64) UInt64Slice {
-	return UInt64Slice{orig: orig}
+func GetUInt64SliceState(ms UInt64Slice) *State {
+	return ms.state
+}
+
+func NewUInt64Slice(orig *[]uint64, state *State) UInt64Slice {
+	return UInt64Slice{orig: orig, state: state}
 }