… .chloggen/pdata-mutation-assertions.yaml
… pdata/internal/generated_wrapper_instrumentationscope.go
… pdata/internal/generated_wrapper_resource.go
… pdata/internal/state.go
… pdata/internal/wrapper_logs.go
… pdata/internal/wrapper_map.go
… pdata/internal/wrapper_metrics.go
… pdata/internal/wrapper_slice.go
… pdata/internal/wrapper_traces.go
… pdata/internal/wrapper_tracestate.go
… pdata/internal/wrapper_value.go
… pdata/pcommon/generated_byteslice.go
… pdata/pcommon/generated_byteslice_test.go
… pdata/pcommon/generated_float64slice.go
… pdata/pcommon/generated_float64slice_test.go
… pdata/pcommon/generated_instrumentationscope.go
… pdata/pcommon/generated_instrumentationscope_test.go
… pdata/pcommon/generated_resource.go
… pdata/pcommon/generated_resource_test.go
… pdata/pcommon/generated_uint64slice.go
… pdata/pcommon/generated_uint64slice_test.go
… pdata/pcommon/map.go
… pdata/pcommon/map_test.go
… pdata/pcommon/slice.go
… pdata/pcommon/slice_test.go
… pdata/pcommon/trace_state.go
… pdata/pcommon/value.go
… pdata/pcommon/value_test.go
… pdata/plog/generated_logrecord.go
… pdata/plog/generated_logrecord_test.go