@dmitryax 2023-09-26
@codeboten the performance should not be affected because the additional field is a part of a struct that stays on the memory stack. I added some benchmarks from my previous work to confirm it:

Before:
```
BenchmarkTracesUsage-10    	    1832	    613221 ns/op	   50631 B/op	    3409 allocs/op
BenchmarkMetricsUsage-10    	     506	   2247812 ns/op	  115778 B/op	    8946 allocs/op
BenchmarkLogsUsage-10    	    5146	    224939 ns/op	   26323 B/op	    1645 allocs/op
```

After:
```
BenchmarkMetricsUsage-10    	     512	   2288719 ns/op	  115777 B/op	    8946 allocs/op
BenchmarkTracesUsage-10    	    1795	    615730 ns/op	   50629 B/op	    3409 allocs/op
BenchmarkLogsUsage-10    	    5469	    217728 ns/op	   26321 B/op	    1645 allocs/op
```