@dmitryax 2023-09-26
Updated PR to add coverage for panics. Had to make some semi-related chores in pdatagen that I moved to another PR https://github.com/open-telemetry/opentelemetry-collector/pull/8537